# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Moco API",
    "description": "Moco API 문서입니다.",
    "version": "1.0.0"
  },
  "host": "localhost:9000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/user/tests": {
      "get": {
        "tags": ["사용자"],
        "summary": "테스트 목록 조회",
        "description": "인증 정보가 필요없는 테스트 목록 조회 \n total: 전체 테스트 수, totalCount: 이용횟수",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "type": "string",
            "description": "NEWEST(최신순), OLDEST(오래된순), MOST_POPULAR(인기순), LEAST_POPULAR(비인기순)"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "success",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 목록을 성공적으로 가져왔습니다.",
                  "type": "string"
                },
                "page": {
                  "example": "number",
                  "type": "number"
                },
                "limit": {
                  "example": "number",
                  "type": "number"
                },
                "total": {
                  "example": "number",
                  "type": "number"
                },
                "result": {
                  "example": [
                    {
                      "id": "number",
                      "name": "string",
                      "subName": "string",
                      "description": "string",
                      "is_published": "boolean",
                      "path": "string",
                      "created_at": "string"
                    }
                  ]
                },
                "totalCount": {
                  "example": "number",
                  "type": "number"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 목록 조회 중 서버 오류가 발생했습니다.",
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/user/tests/{testId}": {
      "get": {
        "tags": ["사용자"],
        "summary": "테스트 상세 조회",
        "description": "테스트 상세 조회(타입, 질문 포함)",
        "parameters": [
          {
            "name": "testId",
            "in": "path",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "success",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 상세 조회를 성공적으로 가져왔습니다.",
                  "type": "string"
                },
                "result": {
                  "test": [
                      {
                          "id": 16,
                          "name": "test3 수정123ㄱㄱ",
                          "description": "test3 설명 수정3454ㄱㄱ",
                          "is_published": false,
                          "path": "tests/62366ebcd5ec59c0737494d9b8c9ed83",
                          "created_at": "2023-12-13T06:10:17.094+00:00",
                          "subName": "test3 서브 타이틀34ㄱㄱ"
                      }
                  ],
                  "types": [
                      {
                          "id": 3,
                          "test_id": 16,
                          "type": "타입1ㅁㄴㅇ",
                          "description": "타입1의 설명ㅂㅈㄷ"
                      },
                      {
                          "id": 4,
                          "test_id": 16,
                          "type": "타입2ㅈㅂ",
                          "description": "타입2의 설명ㄴㅇ"
                      }
                  ],
                  "questions": [
                      {
                          "id": 5,
                          "test_id": 16,
                          "content": "1.질문A3`1",
                          "types": [
                              "닭",
                              "양",
                              "사자",
                              "여우"
                          ]
                      },
                      {
                          "id": 6,
                          "test_id": 16,
                          "content": "1.질문B214124",
                          "types": [
                              "닭",
                              "사자"
                          ]
                      },
                      {
                          "id": 7,
                          "test_id": 16,
                          "content": "2.질문Aㅁㅁㄴㅇ12ttt",
                          "types": [
                              "사자",
                              "늑대"
                          ]
                      },
                      {
                          "id": 8,
                          "test_id": 16,
                          "content": "2.질문B123",
                          "types": [
                              "사자",
                              "개",
                              "닭",
                              "여우"
                          ]
                      }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 ID를 제공해주세요.",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 상세 조회 중 서버 오류가 발생했습니다.",
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/user/testResult/{testId}": {
      "post": {
        "tags": ["사용자"],
        "summary": "테스트 결과 조회",
        "description": "가장 많이 선택된 type 의 row 행을 반환, 동일하게 (ex: [개, 개, 고양이, 사자, 고양이]) 선택될 경우 다중 행 반환 (개, 고양이) \n 결과 조회 후에 /user/saveResult 테스트 결과 저장을 호출 해야합니다.",
        "parameters": [
          {
            "name": "testId",
            "in": "path",
            "type": "string",
            "required": true
          },
          {
            "name": "types",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "properties": {
                "type": {
                  "example": "[]"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "success",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 결과 조회를 성공적으로 가져왔습니다.",
                  "type": "string"
                },
                "result": {
                  "example": {
                    "id": "number",
                    "test_id": "number",
                    "type": "string",
                    "description": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 ID를 제공해주세요. 테스트 타입들을 제공해주세요.",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 결과 조회 중 서버 오류가 발생했습니다.",
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/user/saveResult": {
      "get": {
        "tags": ["사용자"],
        "summary": "테스트 결과 저장",
        "description": "사용자의 테스트 결과 저장 \n /user/getTestResult/:testId 테스트 결과 조회 후에 호출 해야합니다.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "testId": {
                  "example": "number"
                },
                "types": {
                  "example": []
                },
                "descriptions": {
                  "example": []
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "success",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 결과가 성공적으로 저장되었습니다.",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 ID를 제공해주세요. 테스트 결과를 제공해주세요.",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 결과 저장 중 서버 오류가 발생했습니다.",
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/user/statistics": {
      "post": {
        "tags": ["사용자"],
        "summary": "테스트 통계",
        "description": "해당 테스트(id) 의 통계(count) 를 조회 합니다. \n 만약 결과 페이지에서 다른 사람들은 어떤 결과가 나왔는지 통계를 보여주고 싶을때 사용 하면 됩니다.",
        "parameters": [
          {
            "name": "testId",
            "in": "query",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "success",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 통계 조회를 성공적으로 가져왔습니다.",
                  "type": "string"
                },
                "result": {
                  "example": [
                    {
                      "id": "number",
                      "test_id": "number",
                      "name": "string",
                      "type": "string",
                      "description": "string",
                      "count": "number"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 ID를 제공해주세요.",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "example": "error",
                  "type": "string"
                },
                "message": {
                  "example": "테스트 통계 조회 중 서버 오류가 발생했습니다.",
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}